Refactor: Reduce ScheduleService responsibility drift
Schedule response formatting and access validation were mixed into the main orchestration service, making future schedule changes harder to isolate. Move those stable concerns into dedicated collaborators while preserving the existing API contract and transaction shape.

Constraint: User requested only remaining-work item 2 in this pass; transaction boundary changes are intentionally deferred

Rejected: Split generation/query/command services now | broader refactor belongs to later remaining-work items

Confidence: high

Scope-risk: narrow

Directive: Do not change schedule generation transaction boundaries in this commit; handle item 3 separately

Tested: ./gradlew test

Not-tested: Production OAuth/TourAPI/OpenAI integrations
